+++ test/ddns_test.exs
@@ -4,12 +4,12 @@ defmodule Defdo.DDNSTest do
   alias Defdo.Cloudflare.DDNS
 
   test "get config dns records to be monitored" do
-    Application.put_env(:defdo_ddns, Cloudflare, domain_mappings: %{"defdo.de" => ["h"]})
-    assert DDNS.records_to_monitor("defdo.de") == ["defdo.de", "h.defdo.de"]
+    Application.put_env(:defdo_ddns, Cloudflare, domain_mappings: %{"example.com" => ["home"]})
+    assert DDNS.records_to_monitor("example.com") == ["example.com", "home.example.com"]
   end
 
   test "get the cloudflare application key" do
-    Application.put_env(:defdo_ddns, Cloudflare, domain_mappings: %{"defdo.de" => ["h"]})
-    assert DDNS.get_cloudflare_key(:domain_mappings) == %{"defdo.de" => ["h"]}
+    Application.put_env(:defdo_ddns, Cloudflare, domain_mappings: %{"example.com" => ["home"]})
+    assert DDNS.get_cloudflare_key(:domain_mappings) == %{"example.com" => ["home"]}
   end
 end